☀️ 52% – Fix a third sign-extension bug in 080d59b0 (not yet exact)
Same class of bug already fixed on 080dc1ec and 080d82b0: `(bounced +
((u32) bounced >> 31)) >> 1` was compiling to a logical `lsrs` instead
of the reference's arithmetic `asrs`, because the `(u32)` cast forces
the sum unsigned. Cast the sign-bit term back to `(s32)` before adding.
Traced the actual mechanism here more precisely than the prior two
fixes: this compiler's conditional-constant-propagation proves
accumulator==0 from the guarding `if` and forward-substitutes it, so
the wrong promotion produced a stray literal-pool constant (0x7FFFF000
instead of the correct 0xFFFFF000), costing 4 bytes and 2 wrong
instructions.

candidate-show: differing_halfwords 281 -> 277, wrong_instructions
271 -> 268, and candidate size is now byte-count-exact (664 == 664).
Re-verified the trampoline-call fix with --align: both indirect calls
now reproduce the reference's exact stage-to-stack/reload/branch shape.
Re-tried the dominant open residual (work in fp vs r9) with three
declaration-order/scope variants, all no-ops or regressions, confirming
it remains genuinely open. Still not exact.

+++ games/gs1/recon/en/main/080d59b0.c
@@ -167,7 +167,7 @@ s32 Func_080d59b0(void *object)
                             bounced = -(M2C_FIELD(star, s32 *, 16) + 0x2001);
                             M2C_FIELD(star, s32 *, 8) =
                                 M2C_FIELD(star, s32 *, 8) + 4;
-                            bounced = (bounced + ((u32) bounced >> 31)) >> 1;
+                            bounced = (bounced + (s32) ((u32) bounced >> 31)) >> 1;
                             M2C_FIELD(star, s32 *, 4) = 0x5C00000;
                             M2C_FIELD(star, s32 *, 16) = bounced;
                         }
